edit: improved xml parser
