--- src/main/resources/db.properties
@@ -1,3 +1,0 @@
-db.url=jdbc:mysql://localhost:3306/mydatabase
-db.username=myusername
-db.password=mypassword
