--- src/main/resources/person.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<person>
-    <titleId>4</titleId>
-    <fullName>Vernice Metz</fullName>
-    <dateOfBirth>1966-03-15</dateOfBirth>
-    <gender>Female</gender>
-    <address>438 Minh Viaduct</address>
-    <email>vito.harvey@yahoo.com</email>
-    <phoneNumber>1-848-476-4659</phoneNumber>
-</person>
